+++ common/src/main/java/smartin/miapi/client/gui/crafting/crafter/replace/ReplaceView.java
@@ -1,6 +1,5 @@
 package smartin.miapi.client.gui.crafting.crafter.replace;
 
-import com.mojang.blaze3d.systems.RenderSystem;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.Minecraft;
@@ -98,9 +97,6 @@ public class ReplaceView extends InteractAbleWidget {
 
         @Override
         public void renderWidget(GuiGraphics drawContext, int mouseX, int mouseY, float delta) {
-            RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, this.alpha);
-            RenderSystem.enableBlend();
-            RenderSystem.defaultBlendFunc();
             int hoverOffset = 0;
             if (isMouseOver(mouseX, mouseY)) {
                 hoverOffset = 14;