+++ common/src/main/java/smartin/miapi/client/gui/crafting/statdisplay/SingleStatDisplayBoolean.java
@@ -52,8 +52,6 @@ public abstract class SingleStatDisplayBoolean extends SingleStatDisplayDouble {
         double min = Math.min(minValue, Math.min(oldValue, compareToValue));
         double max = Math.max(maxValue, Math.max(oldValue, compareToValue));
 
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, this.alpha);
-
         drawTextureWithEdge(drawContext, texture, getX(), getY(), 339, 6, 51, 19, width, height, 512, 512, 2);
 
         textWidget.setX(this.getX() + 3);