+++ common/src/main/java/smartin/miapi/mixin/enchant/EnchantmentMixin.java
@@ -10,58 +10,26 @@ import net.minecraft.world.item.enchantment.Enchantment;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
-import smartin.miapi.Environment;
-import smartin.miapi.MixinContextFlags;
 import smartin.miapi.item.modular.ModularItem;
-import smartin.miapi.modules.ItemModule;
-import smartin.miapi.modules.ModuleInstance;
 import smartin.miapi.modules.properties.enchanment.AllowedEnchantments;
-import smartin.miapi.modules.properties.util.ComponentApplyProperty;
 
 @Mixin(Enchantment.class)
 public abstract class EnchantmentMixin {
 
     @ModifyReturnValue(method = "isPrimaryItem(Lnet/minecraft/world/item/ItemStack;)Z", at = @At(value = "RETURN"))
     private boolean miapi$adjustPrimaryItem(boolean original, ItemStack itemStack) {
-        if (MixinContextFlags.CALLED_FROM_MUTABLE.get()) {
-            return original;
-        }
-
         if (ModularItem.isModularItem(itemStack)) {
             Enchantment enchantment = (Enchantment) (Object) (this);
-            ModuleInstance moduleInstance = ItemModule.getModules(itemStack);
-            if (moduleInstance != null && moduleInstance.registryAccess != null) {
-                Holder<Enchantment> holder = moduleInstance.registryAccess.registry(Registries.ENCHANTMENT).get().wrapAsHolder(enchantment);
-                if (Environment.isClient() && holder instanceof Holder.Direct<Enchantment>) {
-                    holder = getClient(enchantment);
-                }
-                if (holder != null) {
-                    return AllowedEnchantments.canEnchant(itemStack, holder, original);
-                }
-            }
+            return AllowedEnchantments.canEnchant(itemStack, enchantment, original);
         }
         return original;
     }
 
     @ModifyReturnValue(method = "isSupportedItem", at = @At(value = "RETURN"))
     private boolean miapi$adjustSupportedItem(boolean original, ItemStack itemStack) {
-        if (MixinContextFlags.CALLED_FROM_MUTABLE.get()) {
-            return original;
-        }
-
         if (ModularItem.isModularItem(itemStack)) {
             Enchantment enchantment = (Enchantment) (Object) (this);
-            ModuleInstance moduleInstance = ItemModule.getModules(itemStack);
-            ComponentApplyProperty.trySetup(moduleInstance);
-            if (moduleInstance != null && moduleInstance.registryAccess != null) {
-                Holder<Enchantment> holder = moduleInstance.registryAccess.registry(Registries.ENCHANTMENT).get().wrapAsHolder(enchantment);
-                if (Environment.isClient() && holder instanceof Holder.Direct<Enchantment>) {
-                    holder = getClient(enchantment);
-                }
-                if (holder != null) {
-                    return AllowedEnchantments.isSupported(itemStack, holder, original);
-                }
-            }
+            return AllowedEnchantments.isSupported(itemStack, enchantment, original);
         }
         return original;
     }
@@ -74,21 +42,9 @@ public abstract class EnchantmentMixin {
 
     @ModifyReturnValue(method = "canEnchant(Lnet/minecraft/world/item/ItemStack;)Z", at = @At(value = "RETURN"))
     private boolean miapi$adjustcanEnchant(boolean original, ItemStack itemStack) {
-        if (MixinContextFlags.CALLED_FROM_MUTABLE.get()) {
-            return original;
-        }
         if (ModularItem.isModularItem(itemStack)) {
             Enchantment enchantment = (Enchantment) (Object) (this);
-            ModuleInstance moduleInstance = ItemModule.getModules(itemStack);
-            if (moduleInstance != null && moduleInstance.registryAccess != null) {
-                Holder<Enchantment> holder = moduleInstance.registryAccess.registry(Registries.ENCHANTMENT).get().wrapAsHolder(enchantment);
-                if (Environment.isClient() && holder instanceof Holder.Direct<Enchantment>) {
-                    holder = getClient(enchantment);
-                }
-                if (holder != null) {
-                    return AllowedEnchantments.canEnchant(itemStack, holder, original);
-                }
-            }
+            return AllowedEnchantments.canEnchant(itemStack, enchantment, original);
         }
         return original;
     }