+++ common/src/main/java/smartin/miapi/modules/edit_options/skins/gui/SkinButton.java
@@ -49,11 +49,6 @@ class SkinButton extends InteractAbleWidget implements SkinGui.SortAble {
 
     @Override
     public void renderWidget(GuiGraphics drawContext, int mouseX, int mouseY, float delta) {
-        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, this.alpha);
-        RenderSystem.enableBlend();
-        RenderSystem.defaultBlendFunc();
-        RenderSystem.enableDepthTest();
-        RenderSystem.setShaderTexture(0, skin.textureOptions.texture());
         int hover = this.isMouseOver(mouseX, mouseY) ? skin.textureOptions.ySize() : 0;
         if (!isAllowed) {
             hover = hover * 3;